--- TeamCode/src/main/java/org/firstinspires/ftc/teamcode/UpliftRobot2.java
@@ -1,44 +1,0 @@
-package org.firstinspires.ftc.teamcode;
-
-import com.qualcomm.hardware.bosch.BNO055IMU;
-import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
-import com.qualcomm.robotcore.hardware.DcMotor;
-import com.qualcomm.robotcore.hardware.DcMotorSimple;
-import com.qualcomm.robotcore.hardware.HardwareMap;
-
-public class UpliftRobot2 {
-    public DcMotor leftFront, leftBack, rightFront, rightBack;
-    public float imuAngle;
-    public BNO055IMU imu;
-    public double worldX = 0, worldY = 0, rawAngle = 0, worldAngle = 0;
-
-
-
-    public LinearOpMode opMode;
-    public HardwareMap hardwareMap;
-
-    // robot constructor
-    public UpliftRobot2(LinearOpMode opMode) {
-        this.opMode = opMode;
-        getHardware();
-    }
-
-    public void getHardware() {
-        hardwareMap = opMode.hardwareMap;
-        leftFront = hardwareMap.get(DcMotor.class, "left_front");
-        leftBack = hardwareMap.get(DcMotor.class, "left_back");
-        rightFront = hardwareMap.get(DcMotor.class, "right_front");
-        rightBack = hardwareMap.get(DcMotor.class, "right_back");
-        imu = hardwareMap.get(BNO055IMU.class, "imu");
-        BNO055IMU.Parameters parameters = new BNO055IMU.Parameters();
-        parameters.angleUnit = BNO055IMU.AngleUnit.DEGREES;
-        parameters.accelUnit = BNO055IMU.AccelUnit.METERS_PERSEC_PERSEC;
-        imu.initialize(parameters);
-        leftFront.setDirection(DcMotorSimple.Direction.REVERSE);
-        leftBack.setDirection(DcMotorSimple.Direction.REVERSE);
-        leftFront.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        rightBack.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        leftBack.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-        rightFront.setZeroPowerBehavior(DcMotor.ZeroPowerBehavior.BRAKE);
-    }
-}
